Migrate DataSettings to the composition API
